feat: add restore tests, slight refactor, fix tests
Added a restore test.

Slightly refactor the backup and restore commands to make testing them
easier. Although the restore refactor is kinda jank, it'll work for now.
The reason it's jank is because I'm not sure how to test 'huh'.

Added cleanup for the backup and restore tests.

Fixed tests for the search command.