Config refactor (#5)
* refactor config

This was a pretty major config refactor. This involves returning an
error from decrypting the config so that the user can try as many times
to login. There should be some count, but that's a later point.

This also includes a huge refactor for model.Config to move around
pointers, as that's more idiomatic Go.

* fix: running tests no longer delete main vault