fix: problem with adding and deleting with out of bounds
There was a syncing issue between the filtered and vault lists where if
an item was added and it was the last item in the vault, immediately
after adding it if you wanted to update or delete it, the program would
panic with an out of bounds error.
This syncs the FilteredVault and Vault.

There are also some minor improvements such as q for quitting and
updating the info at the bottom